+++ ElvUI_EltreumUI/Core/Options.lua
@@ -3445,50 +3445,34 @@ function ElvUI_EltreumUI:Configtable()
 	ElvUI_EltreumUI.Options.args.changelog.args.added = ACH:Group(E:TextGradient("Added", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 2)
 	ElvUI_EltreumUI.Options.args.changelog.args.added.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.added.args.description = ACH:Description([[
-Added an option to use the original colors of Mage and Warlock instead of the new ones added in Legion
-Added [name:eltruism:abbrev] and [target:eltruism:abbrev] tags, which accept length args
-Added an option to disable the Gradient Class Bar separately from the Power Gradient
-Added several addon fonts to the font function, such as NameplateSCT and mMediaTag
-Added several shadows to Windtools frames when Windtools shadows is disabled
-Added options to change the Alpha of Health separately from the Backdrop
-Added an option to hide the Event Toaster Frame
+Added an option to show the Stealth Vignette at all times
+Added Path of Twisted Time to the teleport list
 Added more models to the Portrait Fix
-Added a mover to the Battlefield Map
+Added more 1080p optimizations
+Added Loot Method DataText
+Added AvQuest Font
 ]], 3, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.updated = ACH:Group(E:TextGradient("Updated", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.args.description = ACH:Description([[
-Updated Eltruism autoupdate to only update ElvUI settings when using an Eltruism profile
-Updated the Dynamic Datatext option to show on the Datatext tab of Eltruism > Misc
-Updated locales to remove duplicate entries and add a few missing ones
-Updated the LFG skin with a possible fix for the create group taint
-Updated Pet Custom Glow to better handle when the pet is summoned
-Updated Target Combat Icons to be higher resolution on unitframes
-Updated size and position of Player and Target castbars in 1080p
-Updated Quest Item button to move to the correct size in Retail
-Updated quest auto accept to better handle gossip dialogs
-Updated Questie profile to handle the new Ace profiles
-Updated the position of the Addon Compartment frame
-Updated Shadows to better handle other situations
-Updated CVars during install to hide critters
-Updated OmniCD skin to also work in Wrath
-Updated mover for the Addon Compartment
+Updated Ace3 skin to better handle disabled buttons/tabs
+Updated OmniCD skin to be able to be disabled on Classic
 Updated German locale by Dlarge
+Updated Shadows for classbars
+Updated Battlefield Mover
 ]], 5, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed = ACH:Group(E:TextGradient("Fixed", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 4)
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.args.description = ACH:Description([[
-Fixed a possible issue with Leatrix Maps in Classic when attempting to move/adjust the size of the Map
-Fixed a possible issue with free to play accounts when the Subscription Warning shows up
-Fixed an issue with shadows when the orientation of target of target was non default
-Fixed an issue with the Castbar shadows where the orientation was not being checked
-Fixed an issue where the backdrop alpha of unitframes didn't work on group units
-Fixed an issue with the Ace3 skin where it would color item rewards on quests
-Fixed an issue with Portraits where target of target could be missing color
-Fixed an issue with the Ace3 skin where it would flash on the options tree
-Fixed an issue with the Wide transmog when blizzard skins were disabled
-Fixed an issue with [eltruism:healermana] where it would not update
-Fixed an issue with shadows when blizzard skins were disabled
+Fixed an issue with the Raid Warning when ElvUI unitframes were disabled
+Fixed an error with the Quest Skin in Classic when using custom colors
+Fixed an issue where the Portrait Skin could show up on certain frames
+Fixed an issue with the LFG Skin where titles could not be changed
+Fixed an issue where Pet Borders could not be changed in Classic
+Fixed an issue with Shadows when castbar icons were attached
+Fixed an issue with Shadows when using Actionbar backdrops
+Fixed a possible issue with the Quest Auto Accept
+Fixed the Hit Stat calculation in Classic
 ]], 7, "small", nil, nil, nil, nil, "full")
 
 	--[[