+++ Changelog.md
@@ -1,6 +1,27 @@
 ### 🌌 Next Version
 The future is uncertain
 ___
+### 4.0.8
+Added an option to show the Stealth Vignette at all times  
+Added Path of Twisted Time to the teleport list  
+Added more models to the Portrait Fix  
+Added more 1080p optimizations  
+Added Loot Method DataText  
+Added AvQuest Font  
+Updated Ace3 skin to better handle disabled buttons/tabs  
+Updated OmniCD skin to be able to be disabled on Classic  
+Updated German locale by Dlarge  
+Updated Shadows for classbars  
+Updated Battlefield Mover  
+Fixed an issue with the Raid Warning when ElvUI unitframes were disabled  
+Fixed an error with the Quest Skin in Classic when using custom colors  
+Fixed an issue where the Portrait Skin could show up on certain frames  
+Fixed an issue with the LFG Skin where titles could not be changed  
+Fixed an issue where Pet Borders could not be changed in Classic  
+Fixed an issue with Shadows when castbar icons were attached  
+Fixed an issue with Shadows when using Actionbar backdrops  
+Fixed a possible issue with the Quest Auto Accept  
+Fixed the Hit Stat calculation in Classic
 ### 4.0.7
 Added an option to use the original colors of Mage and Warlock instead of the new ones added in Legion  
 Added [name:eltruism:abbrev] and [target:eltruism:abbrev] tags, which accept length args